refactor: clean cron sqlite runtime paths
